Remove loggin wrapper, as no longer necessary
+ Also update missing ZE_RESULT codes from older logger
