+++ source/layers/validation/ze_validation_layer.h
@@ -19,7 +19,7 @@
 #include "zet_entry_points.h"
 #include "zes_entry_points.h"
 #include "zer_entry_points.h"
-#include "logging.h"
+#include "ze_logger.h"
 #include "ze_to_string.h"
 #include "zes_to_string.h"
 #include "zet_to_string.h"
@@ -57,7 +57,7 @@ namespace validation_layer
         std::vector<validationChecker *> validationHandlers;
         std::unique_ptr<HandleLifetimeValidation> handleLifetime;
         
-        std::shared_ptr<loader::Logger> logger;
+        std::shared_ptr<loader::ZeLogger> logger;
 
         static context_t& getInstance() {
             static context_t instance;
