+++ source/utils/CMakeLists.txt
@@ -1,7 +1,7 @@
 # Copyright (C) 2024-2026 Intel Corporation
 # SPDX-License-Identifier: MIT
 
-set(logging_files logging.h logging.cpp ze_logger.h ze_logger.cpp ze_to_string.h zes_to_string.h zet_to_string.h zer_to_string.h)
+set(logging_files ze_logger.h ze_logger.cpp ze_to_string.h zes_to_string.h zet_to_string.h zer_to_string.h)
 add_library(level_zero_utils STATIC ${logging_files})
 
 target_include_directories(level_zero_utils PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
