--- source/utils/logging.h
@@ -1,193 +1,0 @@
-/*
- *
- * Copyright (C) 2024-2026 Intel Corporation
- *
- * SPDX-License-Identifier: MIT
- *
- */
-
-#ifndef level_zero_loader_LOGGING_HPP
-#define level_zero_loader_LOGGING_HPP
-
-#define LOADER_LOG_FILE "ze_loader.log"
-#define LOADER_LOG_FILE_DIRECTORY ".oneapi_logs"
-
-#include <iostream>
-#include <string>
-#include <memory>
-
-#include "ze_api.h"
-#include "ze_util.h"
-#include "ze_logger.h"
-
-#ifdef __linux__
-#include <unistd.h>
-#include <sys/types.h>
-#include <pwd.h>
-#endif // __linux__
-
-namespace loader {
-
-std::string to_string(const ze_result_t result);
-
-enum class Console {
-  out_stdout,
-  out_stderr
-};
-
-class Logger {
-public:
-  // File sink constructor
-  Logger(std::string filename, std::string log_level, bool logging_enabled_env, std::string format = "") {
-    if (logging_enabled_env) {
-      logging_enabled = true;
-      std::string pattern = format.empty() ? _default_pattern : format;
-      LogLevel level = logLevelFromString(log_level);
-      _logger = std::shared_ptr<ZeLogger>(new ZeLogger(filename, level, pattern));
-    }
-  }
-
-  // Console sink constructor
-  Logger(Console out, std::string log_level, bool logging_enabled_env, std::string format = "") {
-    if (logging_enabled_env) {
-      logging_enabled = true;
-      std::string pattern = format.empty() ? _default_pattern : format;
-      LogLevel level = logLevelFromString(log_level);
-      bool use_stderr = (out != Console::out_stdout);
-      _logger = std::shared_ptr<ZeLogger>(new ZeLogger(use_stderr, level, pattern));
-    }
-  }
-
-  ~Logger() {
-    if (!logging_enabled || !_logger)
-      return;
-    _logger->flush();
-  }
-
-  void set_level(LogLevel log_level) {
-    if (!logging_enabled || !_logger)
-      return;
-    _logger->setLevel(log_level);
-  }
-
-  void log_trace(std::string msg) {
-    if (!logging_enabled || !_logger)
-      return;
-    _logger->trace(msg);
-  }
-  void log_debug(std::string msg) {
-    if (!logging_enabled || !_logger)
-      return;
-    _logger->debug(msg);
-  }
-  void log_info(std::string msg) {
-    if (!logging_enabled || !_logger)
-      return;
-    _logger->info(msg);
-  }
-  void log_warning(std::string msg) {
-    if (!logging_enabled || !_logger)
-      return;
-    _logger->warn(msg);
-  }
-  void log_error(std::string msg) {
-    if (!logging_enabled || !_logger)
-      return;
-    _logger->error(msg);
-  }
-  void log_fatal(std::string msg) {
-    if (!logging_enabled || !_logger)
-      return;
-    _logger->critical(msg);
-  }
-
-  void log_performance(std::string msg) {
-    if (!logging_enabled || !_logger)
-      return;
-    _logger->warn("[performance] " + msg);
-  }
-
-bool log_to_console = true;
-bool logging_enabled = false;
-private:
-  static constexpr const char *_default_pattern =
-      "[%Y-%m-%d %H:%M:%S.%e] [thread-id: %t] [%^%l%$] %v";
-
-  std::shared_ptr<ZeLogger> _logger = nullptr;
-};
-
-inline std::shared_ptr<Logger> createLogger() {
-  std::shared_ptr<Logger> zel_logger;
-
-  auto log_directory = getenv_string("ZEL_LOADER_LOG_DIR");
-  if (log_directory.empty()) {
-      std::string home_dir;
-#ifdef _WIN32
-      home_dir = getenv_string("USERPROFILE");
-      if (home_dir == ""){
-          auto home_drive = getenv_string("HOMEDRIVE");
-          auto home_path = getenv_string("HOMEPATH");
-          if ((home_drive != "") && (home_path != "")) {
-              home_dir = home_drive + home_path;
-          } else {
-              home_dir = ".";
-          }
-      }
-      log_directory = home_dir + "\\" + LOADER_LOG_FILE_DIRECTORY;
-#else
-      home_dir = getenv_string("HOME");
-
-      if (home_dir == "") {
-          auto pwdir = getpwuid(getuid())->pw_dir;
-          home_dir = (pwdir == NULL) ? "." : std::string(pwdir);
-      }
-      log_directory = home_dir + "/" + LOADER_LOG_FILE_DIRECTORY;
-#endif
-  }
-  auto loader_file = getenv_string("ZEL_LOADER_LOG_FILE");
-  if (loader_file.empty()){
-      loader_file = LOADER_LOG_FILE;
-  } else {
-      auto log_depr_msg = "ZEL_LOADER_LOG_FILE will be deprecated in a future release";
-      std::cout << log_depr_msg << std::endl;
-  }
-
-  std::string full_log_file_path = "";
-#ifdef _WIN32
-  full_log_file_path = log_directory + "\\" + loader_file;
-#else
-  full_log_file_path = log_directory + "/" + loader_file;
-#endif
-  auto logging_enabled = getenv_tobool( "ZEL_ENABLE_LOADER_LOGGING" );
-  auto log_level = getenv_string("ZEL_LOADER_LOGGING_LEVEL");
-  auto log_console = getenv_tobool("ZEL_LOADER_LOG_CONSOLE");
-
-  if (log_level.empty()) {
-      log_level = "warn";
-  }
-
-  // Default pattern includes thread ID: [timestamp] [thread-id: id] [level] message
-  std::string log_pattern = "[%Y-%m-%d %H:%M:%S.%e] [thread-id: %t] [%^%l%$] %v";
-
-  // Allow users to override the pattern via environment variable
-  auto custom_pattern = getenv_string("ZEL_LOADER_LOG_PATTERN");
-  if (!custom_pattern.empty()) {
-      log_pattern = custom_pattern;
-  }
-
-  if (!log_console) {
-      zel_logger = std::shared_ptr<Logger>(new Logger(full_log_file_path, log_level, logging_enabled, log_pattern));
-  } else {
-      zel_logger = std::shared_ptr<Logger>(new Logger(Console::out_stderr, log_level, logging_enabled, log_pattern));
-  }
-
-  if (!logging_enabled){
-      zel_logger->set_level(LogLevel::off);
-  }
-
-  return zel_logger;
-}
-
-} // namespace loader
-
-#endif
